Revert "Add unit tests for analytics events parsing"
