--- godtoolsTests/Info.plist
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<!--
-   Info.plist
-   godtools
-
-   Created by Ryan Carlson on 6/14/18.
-   Copyright (c) 2018 Cru. All rights reserved.
--->
-<plist version="1.0">
-    <dict>
-    
-    </dict>
-</plist>
